+++ TableViewKit/TableViewManager.swift
@@ -5,7 +5,7 @@ import UIKit
 /// It controls how to display an array of sections, from the header, to its items, to the footer.
 /// It automatically registers any related initial or new views/cells for reusability.
 /// Any changes of the sections will be automatically animated and reflected.
-open class TableViewManager: NSObject {
+open class TableViewManager {
 
     /// The `tableView` linked to this manager
     open let tableView: UITableView
@@ -15,32 +15,26 @@ open class TableViewManager: NSObject {
 
     open var animation: UITableViewRowAnimation = .automatic
 
+    open var dataSource: TableViewKitDataSource? { didSet { tableView.dataSource = dataSource } }
+    open var delegate: TableViewKitDelegate? { didSet { tableView.delegate = delegate } }
+    open var scrollDelegate: UIScrollViewDelegate? { didSet { delegate?.scrollDelegate = scrollDelegate } }
+
     var reusableIdentifiers: Set<String> = []
-
-    /// Initialize a `TableViewManager` with a `tableView`.
-    ///
-    /// - parameter tableView: A `tableView` that will be controlled by the `TableViewManager`
-    public init(tableView: UITableView) {
-        self.tableView = tableView
-        self.sections = []
-        super.init()
-        self.tableView.dataSource = self
-        self.tableView.delegate = self
-        self.setupSections()
-
-    }
 
     /// Initialize a `TableViewManager` with a `tableView` and an initial array of sections
     ///
     /// - parameter tableView: A `tableView` that will be controlled by the `TableViewManager`
     /// - parameter sections: An array of sections
-    public init(tableView: UITableView, sections: [Section]) {
+    public init(tableView: UITableView, sections: [Section] = []) {
         self.tableView = tableView
         self.sections = ObservableArray(array: sections)
-        super.init()
-        self.tableView.dataSource = self
-        self.tableView.delegate = self
+        self.setupDelegates()
         self.setupSections()
+    }
+
+    private func setupDelegates() {
+        self.delegate = TableViewKitDelegate(manager: self)
+        self.dataSource = TableViewKitDataSource(manager: self)
     }
 
     private func setupSections() {
@@ -91,159 +85,8 @@ extension TableViewManager {
             reusableIdentifiers.insert(type.reusableIdentifier)
         }
     }
-}
 
-extension TableViewManager {
-
-    fileprivate func item(at indexPath: IndexPath) -> Item {
+    func item(at indexPath: IndexPath) -> Item {
         return sections[indexPath.section].items[indexPath.row]
-    }
-
-    fileprivate func view(for key: (Section) -> HeaderFooterView, inSection section: Int) -> UIView? {
-        guard case .view(let item) = key(sections[section]) else { return nil }
-
-        let drawer = type(of: item).drawer
-        let view = drawer.view(in: self, with: item)
-        drawer.draw(view, with: item)
-
-        return view
-    }
-
-    fileprivate func title(for key: (Section) -> HeaderFooterView, inSection section: Int) -> String? {
-        if case .title(let value) = key(sections[section]) {
-            return value
-        }
-        return nil
-
-    }
-
-    fileprivate func estimatedHeight(for key: (Section) -> HeaderFooterView, inSection section: Int) -> CGFloat? {
-        let item = key(sections[section])
-        switch item {
-        case .view(let view):
-            guard let height = view.height else { return nil }
-            return height.estimated
-        case .title:
-            return 1.0
-        default:
-            return nil
-        }
-    }
-
-    fileprivate func estimatedHeight(at indexPath: IndexPath) -> CGFloat? {
-        guard let height = item(at: indexPath).height else { return nil }
-        return height.estimated
-    }
-
-    fileprivate func height(for key: (Section) -> HeaderFooterView, inSection section: Int) -> CGFloat? {
-        guard case .view(let view) = key(sections[section]), let value = view.height
-            else { return nil }
-        return value.height
-    }
-
-    fileprivate func height(at indexPath: IndexPath) -> CGFloat? {
-        guard let value = item(at: indexPath).height else { return nil }
-        return value.height
-    }
-
-}
-
-extension TableViewManager: UITableViewDataSource {
-
-    /// Implementation of UITableViewDataSource
-    open func numberOfSections(in tableView: UITableView) -> Int {
-        return sections.count
-    }
-
-    /// Implementation of UITableViewDataSource
-    open func tableView(_ tableView: UITableView, numberOfRowsInSection sectionIndex: Int) -> Int {
-        let section = sections[sectionIndex]
-        return section.items.count
-    }
-
-    /// Implementation of UITableViewDataSource
-    open func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let currentItem = item(at: indexPath)
-        let drawer = type(of: currentItem).drawer
-
-        let cell = drawer.cell(in: self, with: currentItem, for: indexPath)
-        drawer.draw(cell, with: currentItem)
-
-        return cell
-    }
-
-    /// Implementation of UITableViewDataSource
-    open func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
-        return title(for: { $0.header }, inSection: section)
-    }
-
-    /// Implementation of UITableViewDataSource
-    open func tableView(_ tableView: UITableView, titleForFooterInSection section: Int) -> String? {
-        return title(for: { $0.footer }, inSection: section)
-    }
-
-    /// Implementation of UITableViewDataSource
-    // swiftlint:disable:next line_length
-    open func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath) {
-        // Intentionally blank. Required to use UITableViewRowActions
-    }
-}
-
-extension TableViewManager: UITableViewDelegate {
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        guard let currentItem = item(at: indexPath) as? Selectable else { return }
-        currentItem.didSelect()
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return height(at: indexPath) ?? tableView.rowHeight
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        return height(for: { $0.header }, inSection: section) ?? tableView.sectionHeaderHeight
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
-        return height(for: { $0.footer }, inSection: section) ?? tableView.sectionFooterHeight
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, estimatedHeightForRowAt indexPath: IndexPath) -> CGFloat {
-        return estimatedHeight(at: indexPath) ?? tableView.estimatedRowHeight
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, estimatedHeightForHeaderInSection section: Int) -> CGFloat {
-        return estimatedHeight(for: { $0.header }, inSection: section) ?? tableView.estimatedSectionHeaderHeight
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, estimatedHeightForFooterInSection section: Int) -> CGFloat {
-        return estimatedHeight(for: { $0.footer }, inSection: section) ?? tableView.estimatedSectionHeaderHeight
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        return view(for: { $0.header }, inSection: section)
-    }
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, viewForFooterInSection section: Int) -> UIView? {
-        return view(for: { $0.footer }, inSection: section)
-    }
-
-	open func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-		return item(at: indexPath) is Editable
-	}
-
-    /// Implementation of UITableViewDelegate
-    open func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath) -> [UITableViewRowAction]? {
-        guard let item = item(at: indexPath) as? Editable else { return nil }
-        return item.actions
     }
 }